Move the inst_autosetup client to lib/
